requirements.txt
openai == 0.28.0
PyMuPDF == 1.24.0
langchain-community == 0.2.0
Django == 5.0.6
gTTS == 2.5.1
zoomus == 1.2.1
PyMySQL == 1.1.1
python-decouple == 3.8
django-cors-headers == 4.3.1
djangorestframework == 3.15.1
djangorestframework-simplejwt == 5.3.1
django_nextjs == 3.0.0
werkzeug == 3.0.3
fuzzywuzzy == 0.18.0
sentence-transformers == 3.0.1
pandas == 2.2.2
celery == 5.4.0
chromadb == 0.5.0